tools: Move fixture collector to spec
